requirements_new.txt
mysql-connector-python==8.0.28
requests==2.27.1